refactor paths to not clash with common name (libs, artifact_dir, log_dir)
